feat: :sparkles: Adiciona get
