+++ src/main/kotlin/com/tulio/alimentador/model/entity/Animal.kt
@@ -1,4 +1,4 @@
-package com.tulio.model.entity
+package com.tulio.alimentador.model.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
