--- src/main/kotlin/com/tulio/repository/AnimalRepository.kt
@@ -1,16 +1,0 @@
-package com.tulio.repository
-
-import com.tulio.model.entity.Animal
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-
-/**
- * Repositório de animais.
- */
-@Repository
-interface AnimalRepository : CrudRepository<Animal, Long> {
-    /**
-     * Procura o animal pelo id.
-     */
-    fun findAnimalById(identifier: Any)
-}
